// client/src/app/core/models/attributs.ts
export interface Attributs {
  angle: number;
  width: number;
  numberOfSize: number;
  linePattern: string;
  junctionType: string;
  pointJointDiameter: number;
  plotType: string;
  texture: string;
  scaleFactor: number;
  stampChoice: number;
  sideCount: number;
  lineJoint: string;
  alignmentType: string;
  mutator: string;
  font: string;
  fontSize: number;
  minWidth: number;
  eraseSize: number;
  emissionRate: number;
  featherLength: number;
  snappingPoint: string;
  tolerance: number;
  featherWidth: number;
}